+++ src/Stormancer.Plugins/Galaxy/Stormancer.Server.Plugins.Galaxy/GalaxyConstants.cs
@@ -1,0 +1,23 @@
+﻿namespace Stormancer.Server.Plugins.Galaxy
+{
+    /// <summary>
+    /// Galaxy integration constants.
+    /// </summary>
+    public class GalaxyConstants
+    {
+        /// <summary>
+        /// The Galaxy auth provider type.
+        /// </summary>
+        public const string PLATFORM_NAME = "galaxy";
+
+        /// <summary>
+        /// Claim path.
+        /// </summary>
+        public const string GALAXYID_CLAIMPATH = "galaxyId";
+
+        /// <summary>
+        /// Username.
+        /// </summary>
+        public const string USERNAME = "username";
+    }
+}